fix(core): remove three truncating casts, one of them inside a signature
The cast_possible_truncation lint added in the previous commit fired on its
first clean build — the incremental build had cached the affected crates, so the
tree I committed did not pass its own gate. Fixed rather than silenced.

codec.rs wrote its length prefix as `len as u32`. That prefix exists to make the
canonical encoding injective: without it ("ab","c") and ("a","bc") produce
identical bytes. A truncated prefix brings the collision back, inside a
signature, silently. Unreachable in practice, but "unreachable in practice" is a
poor guarantee for something that would corrupt signed data. The prefix is now
u64, which removes the length at which it could happen rather than arguing about
whether that length occurs. Same reasoning for the context length in seal.rs.

olm_type went from `as u8` to a checked conversion returning a wire error. It
cannot fail; if it did it would mean vodozemac changed its contract, and an
error beats a silently wrapped byte.

Both encodings changed, so signatures and sealed boxes from before this commit
do not verify after it. Nothing is released and no data is deployed.

+++ crates/vega-core/src/codec.rs
@@ -28,8 +28,13 @@ impl Writer {
     }
 
     /// Length-prefixed bytes. The prefix is what makes the encoding injective.
+    ///
+    /// The prefix is `u64` rather than `u32` so there is no length at which it
+    /// could truncate. A truncated prefix would make two different field
+    /// sequences encode identically, which is exactly the collision the prefix
+    /// exists to prevent — and it would do so silently, inside a signature.
     pub fn bytes(&mut self, v: &[u8]) -> &mut Self {
-        self.0.extend_from_slice(&(v.len() as u32).to_be_bytes());
+        self.0.extend_from_slice(&(v.len() as u64).to_be_bytes());
         self.0.extend_from_slice(v);
         self
     }
